Fix PhabricatorAuthInviteTestCase::testDuplicateInvite unit test
Summary:
Explicitly cast `$minutes` to int to avoid an error in PHP 8.1 and later.

```
ERROR 8192: Implicit conversion from float to int loses precision at [/var/www/html/phorge/arcanist/src/unit/renderer/ArcanistUnitConsoleRenderer.php:86]
```

Closes T15968

Test Plan: Run `bin/arc unit --everything`.

Reviewers: O1 Blessed Committers, valerio.bozzolan

Reviewed By: O1 Blessed Committers, valerio.bozzolan

Subscribers: tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Maniphest Tasks: T15968

Differential Revision: https://we.phorge.it/D25848

+++ src/unit/renderer/ArcanistUnitConsoleRenderer.php
@@ -82,7 +82,7 @@ final class ArcanistUnitConsoleRenderer extends ArcanistUnitRenderer {
 
   private function formatTime($seconds) {
     if ($seconds >= 60) {
-      $minutes = floor($seconds / 60);
+      $minutes = (int)floor($seconds / 60);
       return pht('%dm%02ds', $minutes, round($seconds % 60));
     }
